requirements.txt
Django>=3.0,<4.0
djangorestframework
django-filter
psycopg2-binary>=2.8
djangorestframework-simplejwt>=4.6.0